requirements.txt
groq
arxiv
pdfplumber
requests
python-dotenv
chromadb
sentence-transformers
networkx
schedule
supabase